fix(计费): 对账页的收入除的是进货折扣不是汇率——碰巧对了三家,另外三家差 7~37 倍
`reconcile.rs` 折人民币时收入和成本都取 `channel_rates.usd_per_cny`。那一列是
「¥1 在那家中转买到多少面值」的**进货折扣**(线上 16 行:0.14/1/1.02/3/5/5.3/7/10),
不是汇率。而用户实际被扣多少人民币只走一条路:`usd_cents_to_wallet_cents`,用后台
那个全局 `usd_per_cny_bps`。供应商给我们打几折,和用户付多少钱没有关系。

实测 7 天,同一批流水两把尺子:
    api.hao.ai              页面 ¥7891.29  实际 ¥7846.48   1.0×  ← 碰巧对
    api.teamorouter.com     页面 ¥ 926.86  实际 ¥ 921.59   1.0×  ← 碰巧对
    polly.modelbridge.cc    页面 ¥ 380.79  实际 ¥2704.48   7.1×
    api.hanhegufei.online   页面 ¥ 335.69  实际 ¥2431.83   7.2×
    api.maomaoai.pro        页面 ¥ 217.80  实际 ¥1546.88   7.1×
    zyz.qingyanzhiying.top  页面 ¥  25.12  实际 ¥ 945.46  37.6×
    合计                    页面 ¥9777.57  实际 ¥16396.86  低画 40%
那两个「碰巧对」是这条错了没被发现的原因:那几家的 0.14 恰好 ≈ 1/7.1023,于是
Claude 线路的数字一直准。成本侧继续用进货折扣——它问的正是那个问题。

原来钉着旧行为的那条断言一并反过来,并写明理由:它的依据「收入的量纲属于线路自带
地址那家」在 c387e33(给钱包加全局折算)之后就不成立了,注释和测试没跟着改。

同批四件:

· **`/api/admin/plan-health` 现在就是 500。** `EXTRACT(epoch FROM …) / 3600.0` 在
  PG 14 起回 numeric,而 Rust 声明的是 f64 → 解码失败。它一直没暴露是因为余额探针
  2026-08-25 才上线,在那之前这条查询恒返回 0 行、走 None 分支。
  顺手把这一整类做成闸门:`scripts/sql-type-check.py` 把每条查询 PREPARE 到真库上
  (不执行不写库),从 pg_prepared_statements 读回列类型和 Rust 声明逐列比。全仓
  119 条查询 / 215 个列位现在全部对得上;去掉那个 cast 就红,变异验过。

· **两条守卫从来没跑过。** `drained_pool_makes_the_gate_reachable_again` 和
  `the_topup_rate_is_per_plan_not_a_single_exchange_rate` 漏了 `#[test]`,在
  `cargo test -- --list` 里一次都不出现——而前者是「用户没钱还能用」那个修复的
  全部行为证据,也就是说那个修复上线时是没有守卫的。
  另有 8 处重复的 `#[test]`(同一个函数注册两遍),就是它吃掉了这两条的属性。
  `scripts/orphan-tests.py` 守这一类。这个脚本第一版是恒真的:判据被本仓测试文档里
  「少了 `#[test]`」那行字喂饱了,剥掉注释才真。

· **免费池部分覆盖时有个双扣口子**(抽干模式上线后是常规路径):池子先出一部分,
  随后事务开启/认领失败 → 整笔按**原始全额**入队 → 恢复重跑 `from_recovery=true`
  跳过免费分支 → 池子付过的那份被向钱包再收一次,而毫点不回滚。改成入队残额。

· **「解析不出单价就静默收 0」现在看得见了。** 判据是 `price_source`:显式配 0
  (deepseek-v4-pro 的 mode:free + {"in":0,"out":0})是运营的决定,安静返回;三样价
  都没配则告警。实测这条路真在漏:grok-4.6 在 2026-08-28 一天 717 次 / 3403 万
  token 收 0(新线路还没填每模型价),deepseek-v4-flash-vision-exp 08-29 是 49/57 次,
  两笔在后台任何一页上都看不见。总览页加一条只在真的有漏时才出现的横幅。

另外给 model_usage 加三列(endpoint_id / wallet_cents / quota_cents)。cost_cents
一个数答不了「这笔钱谁付的」——它把钱包、套餐额度、免费池、以及订阅超配额时运营方
吸收的那段揉在一起,而最后一项此前在任何一张报表上都不存在。不回填历史:旧行的 0
不是「没扣钱」而是「当时没记」,判据用 endpoint_id IS NULL 区分这两件事。

+++ server/src/plan_health.rs
@@ -359,7 +359,7 @@ async fn measured_upstream_per_visible_usd(
            FROM model_usage u, span s WHERE u.created_at BETWEEN s.lo AND s.hi \
          ) \
          SELECT up.usd, q.raw_cents, q.reqs, \
-                EXTRACT(epoch FROM (s.hi - s.lo)) / 3600.0, \
+                (EXTRACT(epoch FROM (s.hi - s.lo)) / 3600.0)::float8, \
                 COALESCE((SELECT sum(drawn) FROM big) / NULLIF((SELECT all_drawn FROM tot), 0), 0) \
          FROM up, q, span s",
     )
